requirements_gpu.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
opencv-python>=4.5.0
Pillow>=8.3.0
pyyaml>=6.0
tqdm>=4.62.0
tensorboard>=2.7.0
scikit-learn>=1.0.0
cupy-cuda11x>=11.0.0